* Refactored some classes. Made some methods const. * Refactored rendering classes. Still work in progress. * Refactored material setup and Scene classes.

+++ trunk/qsculpt/plastilinacore/include/PlastilinaCore/HEdge.h
@@ -33,7 +33,7 @@ public:
     typedef Edge*   weak_ptr;
     typedef Edge*   Ptr;
 
-    typedef unsigned long    size_t;
+    typedef uint    size_t;
     
 private:
     static std::atomic_int NEXT_ID;
